khsmisc: Miscellaneous functions for epidemiology research

This package serves two purposes: (a) loading a set of core packages for epidemiology research and (b) providing a set of custom functions, which also build on the core packages.

Automatically loaded core packages

This package loads a number of useful core packages for epidemiologic data handling as well as descriptive and inferential analyses.

Data handling: the whole universe of the tidyverse-package; variable labels via labelled; reading Excel files via readxl-package.

Tables: formatted tables via gt; descriptive tables via gtsummary-package.

Figures: ggplot; red/green blindness-proof color palettes via scale_color_viridis; arranging figure panels via cowplot (plot_grid).

Analyses: time-to-event analyses using survival; accessing regression model results via broom.

Functions for data handling

varlist: Variable list/dataset inventory, Stata style

write_csv_safely, save_safely, pdf_safely: Export CSV files, R objects, or PDFs without overwriting existing files

Functions for vectors

scale_to_range: Scale a vector to a given range

Mode: Most common value

roundp: Round and format p-values

Functions for descriptive tables

mytab: A contingency table for categorical variables

tsummary: A descriptive table for continuous variables

mygt: Format and print any tibble as a 'gt' table

table1: Formatted stratified descriptive tables using gtsummary

table2: Stratified result tables

rates: Event counts, person-time, incidence rates, and cumulative incidence

Functions for plots

stripplot: Box-whiskers/dot plot

corrmat: Correlation matrix

exclusion_flowchart: Flowcharts of exclusions from study

See also