Event Counts, Person-Time, Incidence Rates, and Cumulative Incidence (Risk).

Summary table of overall or stratified incidence rates and cumulative indicence. To obtain stratified results, group the tibble before calling rates. See examples.

rates(
  data,
  time,
  event,
  conversion = 1,
  rate_digits = 1,
  risk_digits = 1,
  factor = 1000
)

Arguments

data
Data frame/tibble. If grouped using group_by, then stratified results will be returned.

time
Variable indicating follow-up time.

event
Variable indicating events (event = 1, censor = 0).

conversion
Divisor for follow-up time. Defaults to 1. Set conversion = 12 if time is in months and rates per person-years are desired.

rate_digits
Decimal digits to display for rates. Defaults to 1.

risk_digits
Decimal digits to display for cumulative incidence (risk). Defaults to 1.

factor
Person-time units for the denominator. Defaults to 1000, e.g., for rates per 1000 person-years of follow-up.

Value

Tibble:

n Count of participants

Events Count of events

Person-time Sum of person-time

Rate per person-time Incidence rate

Risk (Kaplan-Meier, %) Cumulative incidence at the end of follow-up, with log-log 95% confidence intervals

References

Greenland S, Rothman KJ (2008). Introduction to Categorical Statistics. In: Rothman KJ, Greenland S, Lash TL. Modern Epidemiology, 3rd edition. Philadelpha, PA: Lippincott Williams & Wilkins. Page 242. (Poisson/large-sample approximation for variance of incidence rates)

Examples

data(ovarian)  # from survival package

# Unstratified
ovarian %>%
  rates(time = "futime", event = "fustat")

#> # A tibble: 1 x 5
#>       n Events `Person-time` `Rate per person-time` `Risk (Kaplan-Meier, %)`
#>   <int>  <dbl>         <dbl> <chr>                  <chr>                   
#> 1    26     12         15588 0.8 (0.4-1.4)          50.3 (32.1-71.8)        

# Convert time to person-years
ovarian %>%
  rates(time = "futime", event = "fustat", conversion = 12)

#> # A tibble: 1 x 5
#>       n Events `Person-time` `Rate per person-time` `Risk (Kaplan-Meier, %)`
#>   <int>  <dbl>         <dbl> <chr>                  <chr>                   
#> 1    26     12          1299 9.2 (5.2-16.3)         50.3 (32.1-71.8)        

# Stratified by "ecog.ps":
ovarian %>%
  group_by(ecog.ps) %>%
  rates(time = "futime", event = "fustat", conversion = 12)

#> # A tibble: 2 x 6
#>   ecog.ps     n Events `Person-time` `Rate per person-tim… `Risk (Kaplan-Meier,…
#>     <dbl> <int>  <dbl>         <dbl> <chr>                 <chr>                
#> 1       1    14      5           694 7.2 (3.0-17.3)        37.5 (17.4-68.5)     
#> 2       2    12      7           605 11.6 (5.5-24.3)       62.5 (36.2-88.3)