Alternative styling for gt tables exported as kable

This function can be called on a gt table and will function just as tab_footnote. If called on a kable table, it will print the kable table, ignoring the styling.

tab_style(data, style, locations)

Arguments

data

gt table or kable table

style

Style definition, e.g., per cell_text. Will be ignored for a kable table.

locations

Location, as per gt, e.g., cells_body. Will be ignored for a kable table.

Value

gt table or kable table, whichever format data is in.