Stratified Descriptive Tables Using gtsummary

table1 wraps tbl_summary with some custom options:

Condensed formatting for better space use.

Stratification in columns with informative spanning headers, plus an "overall" column if desired.

First row with total count.

Rounding of continuous values to one decimal digit by default.

table1(
  data,
  ...,
  by = NULL,
  overall = TRUE,
  label = "",
  digits = list(all_continuous() ~ c(1, 1)),
  statistic = NULL
)

Arguments

data
Data frame/tibble to print. Required.

...
Optional: Rows to include in table. If none are provided, all variables will be shown. Supports tidy evaluation; see examples.

by
Optional: Stratification variable for columns.

overall
Optional: Show "Overall" column if stratifying? Defaults to TRUE.

label
Optional: Label for strata. If empty, the variable label or variable name of the by variable will be used.

digits
Optional: Level of precision (decimal digits) for statistics. By default, all continuous variables are shown with one decimal digits for each statistic: all_continuous() ~ c(1, 1). Precision can also be changed for individual variables, e.g.: all_continuous() ~ c(1, 1), a_specific_variable ~ c(2, 2)

statistic
Optional: Summary statistic to the shown for specific variables, e.g.: list(a_specific_variable ~ "{min}, {max}"). Passed on to tbl_summary(statistic = ...).

Value

Formatted table. Continuous variables are median (quartile 1, quartile 3); categorical variables are count (column percent).

Examples

data(mtcars)

# Stratify by "gear":
mtcars %>%
  table1(by = gear)

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #yrfwjgkqpp .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #yrfwjgkqpp .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #yrfwjgkqpp .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #yrfwjgkqpp .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 4px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #yrfwjgkqpp .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #yrfwjgkqpp .gt_col_headings {
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #yrfwjgkqpp .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: bold;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #yrfwjgkqpp .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: bold;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #yrfwjgkqpp .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #yrfwjgkqpp .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #yrfwjgkqpp .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #yrfwjgkqpp .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #yrfwjgkqpp .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #yrfwjgkqpp .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #yrfwjgkqpp .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #yrfwjgkqpp .gt_row {
#>   padding-top: 0px;
#>   padding-bottom: 0px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #yrfwjgkqpp .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #yrfwjgkqpp .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #yrfwjgkqpp .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #yrfwjgkqpp .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #yrfwjgkqpp .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #yrfwjgkqpp .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #yrfwjgkqpp .gt_table_body {
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #yrfwjgkqpp .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #yrfwjgkqpp .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #yrfwjgkqpp .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #yrfwjgkqpp .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #yrfwjgkqpp .gt_left {
#>   text-align: left;
#> }
#> 
#> #yrfwjgkqpp .gt_center {
#>   text-align: center;
#> }
#> 
#> #yrfwjgkqpp .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #yrfwjgkqpp .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #yrfwjgkqpp .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #yrfwjgkqpp .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #yrfwjgkqpp .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #yrfwjgkqpp .gt_footnote_marks {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="yrfwjgkqpp" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;"><table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_center gt_columns_bottom_border" rowspan="2" colspan="1"><strong>Characteristic</strong></th>
#>       <th class="gt_col_heading gt_center gt_columns_bottom_border" rowspan="2" colspan="1"><strong>Overall</strong></th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="3">
#>         <span class="gt_column_spanner"><strong>By gear</strong></span>
#>       </th>
#>     </tr>
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1"><strong>3</strong></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1"><strong>4</strong></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1"><strong>5</strong></th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">N</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">32</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">15</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">12</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">5</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">mpg</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">19.2 (15.4, 22.8)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">15.5 (14.5, 18.4)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">22.8 (21.0, 28.1)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">19.7 (15.8, 26.0)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">cyl</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">4</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">11 (34%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">1 (7%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">8 (67%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">2 (40%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">6</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">7 (22%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">2 (13%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">4 (33%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">1 (20%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">8</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">14 (44%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">12 (80%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">0 (0%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">2 (40%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">disp</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">196.3 (120.8, 326.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">318.0 (275.8, 380.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">130.9 (78.9, 160.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">145.0 (120.3, 301.0)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">hp</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">123.0 (96.5, 180.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">180.0 (150.0, 210.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">94.0 (65.8, 110.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">175.0 (113.0, 264.0)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">drat</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">3.7 (3.1, 3.9)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">3.1 (3.0, 3.2)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">3.9 (3.9, 4.1)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">3.8 (3.6, 4.2)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">wt</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">3.3 (2.6, 3.6)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">3.7 (3.5, 4.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">2.7 (2.1, 3.2)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">2.8 (2.1, 3.2)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">qsec</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">17.7 (16.9, 18.9)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">17.4 (17.0, 18.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">18.8 (18.5, 19.6)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">15.5 (14.6, 16.7)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">vs</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">14 (44%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">3 (20%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">10 (83%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">1 (20%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">am</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">13 (41%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">0 (0%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">8 (67%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">5 (100%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">carb</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">1</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">7 (22%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">3 (20%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">4 (33%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">0 (0%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">2</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">10 (31%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">4 (27%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">4 (33%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">2 (40%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">3</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">3 (9%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">3 (20%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">0 (0%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">0 (0%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">4</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">10 (31%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">5 (33%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">4 (33%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">1 (20%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">6</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">1 (3%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">0 (0%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">0 (0%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">1 (20%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">8</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">1 (3%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">0 (0%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">0 (0%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">1 (20%)</td>
#>     </tr>
#>   </tbody>
#>   
#>   
#> </table></div>

# Show only selected variables in rows:
mtcars %>%
  table1(cyl, mpg, disp, by = gear)

#> <style>html {
#>   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Helvetica Neue', 'Fira Sans', 'Droid Sans', Arial, sans-serif;
#> }
#> 
#> #nscsdqhqpm .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #nscsdqhqpm .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #nscsdqhqpm .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #nscsdqhqpm .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 0;
#>   padding-bottom: 4px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #nscsdqhqpm .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #nscsdqhqpm .gt_col_headings {
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #nscsdqhqpm .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: bold;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #nscsdqhqpm .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: bold;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #nscsdqhqpm .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #nscsdqhqpm .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #nscsdqhqpm .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #nscsdqhqpm .gt_group_heading {
#>   padding: 8px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #nscsdqhqpm .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #nscsdqhqpm .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #nscsdqhqpm .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #nscsdqhqpm .gt_row {
#>   padding-top: 0px;
#>   padding-bottom: 0px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #nscsdqhqpm .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 12px;
#> }
#> 
#> #nscsdqhqpm .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #nscsdqhqpm .gt_first_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #nscsdqhqpm .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #nscsdqhqpm .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #nscsdqhqpm .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #nscsdqhqpm .gt_table_body {
#>   border-top-style: none;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #nscsdqhqpm .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #nscsdqhqpm .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #nscsdqhqpm .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #nscsdqhqpm .gt_sourcenote {
#>   font-size: 90%;
#>   padding: 4px;
#> }
#> 
#> #nscsdqhqpm .gt_left {
#>   text-align: left;
#> }
#> 
#> #nscsdqhqpm .gt_center {
#>   text-align: center;
#> }
#> 
#> #nscsdqhqpm .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #nscsdqhqpm .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #nscsdqhqpm .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #nscsdqhqpm .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #nscsdqhqpm .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #nscsdqhqpm .gt_footnote_marks {
#>   font-style: italic;
#>   font-size: 65%;
#> }
#> </style>
#> <div id="nscsdqhqpm" style="overflow-x:auto;overflow-y:auto;width:auto;height:auto;"><table class="gt_table">
#>   
#>   <thead class="gt_col_headings">
#>     <tr>
#>       <th class="gt_col_heading gt_center gt_columns_bottom_border" rowspan="2" colspan="1"><strong>Characteristic</strong></th>
#>       <th class="gt_col_heading gt_center gt_columns_bottom_border" rowspan="2" colspan="1"><strong>Overall</strong></th>
#>       <th class="gt_center gt_columns_top_border gt_column_spanner_outer" rowspan="1" colspan="3">
#>         <span class="gt_column_spanner"><strong>By gear</strong></span>
#>       </th>
#>     </tr>
#>     <tr>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1"><strong>3</strong></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1"><strong>4</strong></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1"><strong>5</strong></th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">N</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">32</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">15</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">12</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">5</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">cyl</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;"></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">4</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">11 (34%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">1 (7%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">8 (67%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">2 (40%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">6</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">7 (22%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">2 (13%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">4 (33%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">1 (20%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="text-align: left; text-indent: 10px; vertical-align: top;">8</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">14 (44%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">12 (80%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">0 (0%)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">2 (40%)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">mpg</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">19.2 (15.4, 22.8)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">15.5 (14.5, 18.4)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">22.8 (21.0, 28.1)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">19.7 (15.8, 26.0)</td>
#>     </tr>
#>     <tr>
#>       <td class="gt_row gt_left" style="font-weight: bold; text-align: left; vertical-align: top;">disp</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">196.3 (120.8, 326.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">318.0 (275.8, 380.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">130.9 (78.9, 160.0)</td>
#>       <td class="gt_row gt_center" style="text-align: left; vertical-align: top;">145.0 (120.3, 301.0)</td>
#>     </tr>
#>   </tbody>
#>   
#>   
#> </table></div>