Table 2: Counts and Cox regression results

This function obtains descriptive and inferential results for survival analyses in the format of a table stratified by exposure and, if requested, by effect modifiers.

table2(design, data, layout = "rows", factor = 1000)

Arguments

design
Data frame that sets up the table. See Details.

data
Dataset to be used for all analyses.

layout
"rows" uses the design as rows and exposure categories as columns. "cols" is the opposite: design as columns and exposure categories as rows.

factor
Used for type = "rates": Factor to multiply events per person-time by. Defaults to 1000.

Value

Tibble. Get formatted output by passing on to mygt.

Details

The main input parameter is the dataset design, which must consist of the following variables:

label A label for each row (or column).

time The time variable of the survival model.

event The event variable of the survival model. Events must be 1, censored observations 0.

exposure The exposure variable. Must be categorical (factor or logical).

effect_modifier An effect modifier variable. Use NULL or NA to leave blank.

stratum A stratum of the effect modifier. Use NULL to leave blank. (NA will evaluate observations with missing effect modifier.)

confounders A string in the format "+ confounder1 + confounder2" that will be substituted into into coxph(formula = exposure + confounders). Use "" (empty string) to leave blank. Can add "+ strata(site)" to obtain Cox models with stratification by, e.g., site.

type The statistic requested:

"cox" Hazard ratio and 95% confidence interval. Models will be fit separately for each stratum of the effect_modifier.

"cox" Hazard ratio and 95% confidence interval. A 'joint' model for exposure and effect modifier with a single reference category will be fit.

"events" Event count

"time" Person-time

"total" Number of observations

"events/time" Events slash person-time

"events/total" Events slash number of observations

"rate" Event rate: event count divided by person-time

"blank" An empty line

Hint: Use tibble, tribble, and mutate to construct the design dataset, especially variables that are used repeatedly (e.g., time, event). See examples.

Examples

# Load 'ovarian' dataset from survival package
data(ovarian)

# The exposure (here, 'rx') must be categorical
ovarian <- ovarian %>%
  as_tibble() %>%
  mutate(rx = factor(rx),
         futime = futime / 365.25)  # presumably years

# Set design elements that vary by row
tribble(
   ~label,                      ~stratum, ~confounders, ~type,
  "Overall: Events",            NULL,     "",           "events",
  "Overall: Person-years",       NULL,     "",           "time",
  "Overall: Rate/1000 years",    NULL,     "",           "rate",
  "  Unadjusted HR (95% CI)",   NULL,     "",           "cox",
  "  Age-adjusted HR (95% CI)", NULL,     "+ age",      "cox",
  "",                           NULL,     "",           "blank",
  "Stratified models",          NULL,     "",           "blank",
  "ECOG PS1 (events/N)",        1,        "",           "events/total",
  "  Unadjusted",               1,        "",           "cox",
  "  Age-adjusted",             1,        "+ age",      "cox",
  "ECOG PS2 (events/N)",        2,        "",           "events/total",
  "  Unadjusted",               2,        "",           "cox",
  "  Age-adjusted",             2,        "+ age",      "cox",
  "",                           NULL,     "",           "blank",
  "Joint model, age-adj.",      NULL,     "",           "blank",
  "  ECOG PS1",                 1,        "+ age",      "cox_joint",
  "  ECOG PS2",                 2,        "+ age",      "cox_joint") %>%
# Set design elements that are the same for all rows
  mutate(exposure = "rx", event = "fustat", time = "futime",
        effect_modifier = "ecog.ps") %>%
# Generate table2
  table2(data = ovarian)

#> # A tibble: 17 x 3
#>    rx                           `1`                `2`               
#>    <chr>                        <chr>              <chr>             
#>  1 "Overall: Events"             NA                 NA               
#>  2 "Overall: Person-years"      "18"               "24"              
#>  3 "Overall: Rate/1000 years"   "538.5"            "384.6"           
#>  4 "  Unadjusted HR (95% CI)"   "1 (reference)"    "0.55 (0.17-1.74)"
#>  5 "  Age-adjusted HR (95% CI)" "1 (reference)"    "0.45 (0.13-1.54)"
#>  6 ""                           ""                 ""                
#>  7 "Stratified models"          ""                 ""                
#>  8 "ECOG PS1 (events/N)"        "4/7"              "1/7"             
#>  9 "  Unadjusted"               "1 (reference)"    "0.20 (0.02-1.80)"
#> 10 "  Age-adjusted"             "1 (reference)"    "0.21 (0.02-2.02)"
#> 11 "ECOG PS2 (events/N)"        "3/6"              "4/6"             
#> 12 "  Unadjusted"               "1 (reference)"    "1.16 (0.26-5.28)"
#> 13 "  Age-adjusted"             "1 (reference)"    "0.62 (0.10-3.64)"
#> 14 ""                           ""                 ""                
#> 15 "Joint model, age-adj."      ""                 ""                
#> 16 "  ECOG PS1"                 "1 (reference)"    "0.18 (0.02-1.72)"
#> 17 "  ECOG PS2"                 "0.63 (0.13-3.09)" "0.48 (0.11-2.06)"

# Use '%>% mygt()' to get formatted version (not shown)