Events, Person-Time, Incidence Rates, and Cumulative Incidence (Risk)

Summary table of overall or stratified incidence rates and cumulative incidence.

rates(
  data,
  time,
  event,
  by,
  risk_time = NULL,
  conversion = 1,
  rate_digits = 1,
  risk_digits = 1,
  factor = 1000
)

Arguments

data

Data frame/tibble. If grouped using group_by, then stratified results will be returned.

time

Variable indicating follow-up time.

event

Variable indicating events (event = 1, censor = 0).

by

Optional. Variable to stratify results by. Defaults to how the dataset is grouped, i.e., if it is grouped already, results will be shown by the stratification variable.

risk_time

Optional. Time point at which cumulative incidence (risk) should be calculated. Provide in original units of follow-up time, i.e., before potential division by the conversion factor. If not provided, defaults to the end of follow-up, which may not be the best interpretable timepoint.

conversion

Optional. Divisor for follow-up time. Defaults to 1. Set conversion = 12 if time is in months and rates per person-years are desired.

rate_digits

Optional. Decimal digits to display for rates. Defaults to 1.

risk_digits

Optional. Decimal digits to display for cumulative incidence (risk). Defaults to 1.

factor

Optional. Person-time units for the denominator. Defaults to 1000, e.g., for rates per 1000 person-years of follow-up.

Value

Tibble:

n Count of participants.

Events Count of events.

Person-time Sum of person-time.

Rate per person-time Incidence rate.

Risk (%) Cumulative incidence at time risk_time or, if not provided, at the end of follow-up, with log-log 95% confidence intervals.

Details

This function is not being developed further. All functionality has been implemented in table2.

References

Greenland S, Rothman KJ (2008). Introduction to Categorical Statistics. In: Rothman KJ, Greenland S, Lash TL. Modern Epidemiology, 3rd edition. Philadelpha, PA: Lippincott Williams & Wilkins. Page 242. (Poisson/large-sample approximation for variance of incidence rates)

Examples

data(cancer, package = "survival")
cancer <- cancer %>%
  dplyr::mutate(status = status - 1) %>%
  dplyr::filter(ph.ecog < 3)

# Unstratified
cancer %>%
  rates(time = time, event = status)
#> # A tibble: 1 × 5
#>       n Events `Person-time` `Rate per person-time` `Risk (%)`      
#>   <int>  <dbl>         <dbl> <chr>                  <chr>           
#> 1   226    163         69404 2.3 (2.0-2.7)          94.9 (89.0-98.2)

# Show risk (cumulative incidence) at 1 year,
# show follow-up time to person-years
cancer %>%
  rates(time = time, event = status,
        risk_time = 1 * 365.25, conversion = 365.25)
#> # A tibble: 1 × 5
#>       n Events `Person-time` `Rate per person-time` `Risk (%)`      
#>   <int>  <dbl>         <dbl> <chr>                  <chr>           
#> 1   226    163           190 857.8 (735.7-1000.1)   58.7 (51.8-65.8)

# Stratified by "ph.ecog" and "sex":
cancer %>%
  rates(time = time, event = status,
        by = c(ph.ecog, sex),
        risk_time = 1 * 365.25, conversion = 365.25)
#> # A tibble: 6 × 7
#> # Groups:   ph.ecog [3]
#>   ph.ecog   sex     n Events `Person-time` `Rate per person-time` `Risk (%)`    
#>     <dbl> <dbl> <int>  <dbl>         <dbl> <chr>                  <chr>         
#> 1       0     1    36     28            36 775.7 (535.6-1123.5)   52.0 (36.0-70…
#> 2       0     2    27      9            25 365.9 (190.4-703.2)    38.2 (19.0-66…
#> 3       1     1    71     54            53 1015.9 (778.1-1326.5)  67.4 (55.2-79…
#> 4       1     2    42     28            44 634.5 (438.1-919.0)    39.9 (26.1-57…
#> 5       2     1    29     28            17 1623.6 (1121.0-2351.5) 79.9 (63.7-92…
#> 6       2     2    21     16            15 1081.2 (662.4-1764.9)  76.4 (55.3-92…