Table 1: Stratified Descriptive Tables

table1 wraps tbl_summary with some custom options:

Condensed formatting for better space use.

Stratification in columns with informative spanning headers, plus an "overall" column if desired.

First row with total count.

Rounding of continuous values to one decimal digit by default.

table1(
  data,
  ...,
  by = NULL,
  overall = TRUE,
  label = "",
  digits = list(all_continuous() ~ c(1, 1)),
  statistic = NULL
)

Arguments

data
Data frame/tibble to print. Required.

...
Optional: Rows to include in table. If none are provided, all variables will be shown. Supports tidy evaluation; see examples.

by
Optional: Stratification variable for columns.

overall
Optional: Show "Overall" column if stratifying? Defaults to TRUE.

label
Optional: Label for strata. If empty, the variable label or variable name of the by variable will be used.

digits
Optional: Level of precision (decimal digits) for statistics. By default, all continuous variables are shown with one decimal digits for each statistic: all_continuous() ~ c(1, 1). Precision can also be changed for individual variables, e.g.: all_continuous() ~ c(1, 1), a_specific_variable ~ c(2, 2)

statistic
Optional: Summary statistic to the shown for specific variables, e.g.: list(a_specific_variable ~ "{min}, {max}"). Passed on to tbl_summary(statistic = ...).

Value

Formatted table. Continuous variables are median (quartile 1, quartile 3); categorical variables are count (column percent).

Examples

data(mtcars)

# Stratify by "gear":
if (FALSE) {
mtcars %>%
  table1(by = gear)
}

# Show only selected variables in rows:
if (FALSE) {
mtcars %>%
  table1(cyl, mpg, disp, by = gear)
}