Table 2: Stratified Result Tables

This function displays descriptive and inferential results for binary, continuous, and survival data in the format of a table stratified by exposure and, if requested, by effect modifiers.

This function is intended only for tabulations of final results. Model diagnostics for regression models need to be conducted separately.

table2(
  design,
  data,
  layout = "rows",
  factor = 1000,
  risk_percent = FALSE,
  diff_digits = dplyr::if_else(risk_percent == TRUE, true = 0, false = 2),
  ratio_digits = 2,
  rate_digits = 1,
  to = NULL
)

Arguments

design
Design matrix (data frame) that sets up the table. See Details.

data
Dataset to be used for all analyses.

layout
Optional. "rows" uses the design as rows and exposure categories as columns. "cols" is the opposite: design as columns and exposure categories as rows. Defaults to "rows".

factor
Optional. Used for type = "rates": Factor to multiply events per person-time by. Defaults to 1000.

risk_percent
Optional. Show risk and risk difference estimates in percentage points instead of proportions. Defaults to FALSE.

diff_digits
Optional. Number of decimal digits to show for rounding of means, risks, and risk difference estimates. Defaults to 2 for risk_percent = FALSE and to 0 for risk_percent = TRUE. Can override for each line in type.

ratio_digits
Optional. Number of decimal digits to show for ratio estimates. Defaults to 2. Can override for each line in type.

rate_digits
Optional. Number of decimal digits to show for rates. Defaults to 1. Can override for each line in type.

to
Optional. Separator between the lower and the upper bound of the 95% confidence interval (and interquartile range for medians). Defaults to " to " for means, medians, and mean differences; defaults to "-" otherwise.

Value

Tibble. Get formatted output by passing on to mygt.

Details

The main input parameter is the dataset design. Always required are the columns label, type, and exposure, as well as outcome for binary outcomes or time and event for survival outcomes:

label A label for each row (or column).

time The time variable for survival data. Needed for, e.g., type = "hr" and type = "rate" (i.e., whenever outcome is not used).

time2 The second time variable for late entry models. Only used in conjunction with time. If provided, time will become the entry time and time2 the exit time, following conventions of Surv.

event The event variable for survival data. Events are typically 1, censored observations 0. Needed for, e.g., type = "hr" and type = "rate" (i.e., whenever outcome is not used).

outcome The outcome variable for non-survival data (i.e., whenever event and time are not used). For risk/prevalence data, this variable must be 0/1 or FALSE/TRUE.

exposure The exposure variable. Must be categorical (factor or logical).

effect_modifier Optional. A categorical effect modifier variable. Use NULL or NA to leave blank.

stratum Optional. A stratum of the effect modifier. Use NULL to leave blank. NA will evaluate observations with missing data for the effect_modifier.

confounders Optional. A string in the format "+ var1 + var2" that will be substituted into into formula = exposure + confounders. Use "" (empty string) to leave blank; the default. For Cox models, can add "+ strata(site)" to obtain models with stratification by, e.g., site. For Poisson models, can add "+ offset(log(persontime))" to define, e.g., persontime as the offset.

type The statistic requested (case-insensitive):

Comparative estimates from regression models with 95% confidence intervals:

"hr" Hazard ratio from Cox proportional hazards regression.

"irr" Incidence rate ratio for count outcomes from Poisson regression model.

"rr" Risk ratio (or prevalence ratio) from riskratio. Can request specific model fitting approach and, for marginal standardization only, the number of bootstrap repeats. Examples: "rrglm_start" or "rrmargstd2000". (Note absence of white space.)

"rd" Risk difference (or prevalence difference) from riskdiff. Can request model fitting approach and bootstrap repeats as for "rr".

"diff" Mean difference from linear model.

"quantreg" Quantile difference from quantile regression using rq with method = "fn". By default, this is the difference in medians. For a different quantile, e.g., the 75th percentile, use "quantreg0.75". Note absence of white space before the quantile.

"fold" Fold change from generalized linear model with log link (i.e., ratio of arithmetic means).

"foldlog" Fold change from linear model after log transformation of the outcome (i.e., ratio of geometric means).

"or" Odds ratio from logistic regression.

Absolute estimates per exposure category:

"events" Event count.

"time" Person-time.

"outcomes" Outcome count.

"total" Number of observations.

"events/time" Events slash person-time.

"events/total" Events slash number of observations.

"cases/controls" Cases and non-cases; for case-control studies.

"risk" Risk (or prevalence), i.e., events divided by number of observations. Change between display as proportion or percent using the parameter risk_percent.

"risk (ci)" Risk with 95% confidence interval (Wilson score interval for binomial proportions, see scoreci).

"rate" Event rate: event count divided by person-time, multiplied by factor.

"rate (ci)" Event rate with 95% confidence interval (Poisson-type interval, see rates).

"outcomes (risk)" A combination: Outcomes followed by risk in parentheses.

"outcomes/total (risk)" A combination: Outcomes slash total followed by risk in parentheses.

"events/time (rate)" A combination: Events slash time followed by rate in parentheses.

"mean" Mean.

"mean (ci)" Mean and 95% CI.

"median" Median.

"median (iqr)" Median and interquartile range.

"blank" or "" An empty line.

By default, regression models will be fit separately for each stratum of the effect_modifier. Append "_joint" to "hr", "rr", "rd", "irr", "diff", "fold", "foldlog", "quantreg", or "or" to obtain "joint" models for exposure and effect modifier that have a single reference category. Example: type = "hr_joint". The reference categories for exposure and effect modifier are their first factor levels, which can be changed using fct_relevel.

Digits for rounding estimates can be specified for each line separately. Example: type = "diff (ci) 3" to request a mean difference and its 95% CI rounded to 3 decimal digits (note the space before 3).

Hint: Use tibble, tribble, and mutate to construct the design dataset, especially variables that are used repeatedly (e.g., exposure, time, event, or outcome). See examples.

Examples

# Load 'cancer' dataset from survival package
data(cancer, package = "survival")

# The exposure (here, 'sex') must be categorical
cancer <- cancer %>%
  tibble::as_tibble() %>%
  dplyr::mutate(sex = factor(sex, levels = 1:2,
                             labels = c("Men", "Women")),
                time = time / 365.25,
                status = status - 1)

# Example 1: Binary outcomes (use 'outcome' variable)
# Set table design
design1 <- tibble::tibble(
  label = c("Outcomes",
            "Total",
            "Outcomes/Total",
            "Risk",
            "Risk (CI)",
            "Outcomes (Risk)",
            "Outcomes/Total (Risk)",
            "RR",
            "RD")) %>%
  dplyr::mutate(type = label,
                exposure = "sex",
                outcome = "status")

# Generate table2
table2(design = design1, data = cancer)

#> # A tibble: 9 x 3
#>   sex                   Men              Women                 
#>   <chr>                 <chr>            <chr>                 
#> 1 Outcomes              112              53                    
#> 2 Total                 138              90                    
#> 3 Outcomes/Total        112/138          53/90                 
#> 4 Risk                  0.81             0.59                  
#> 5 Risk (CI)             0.81 (0.74-0.87) 0.59 (0.49-0.68)      
#> 6 Outcomes (Risk)       112 (0.81)       53 (0.59)             
#> 7 Outcomes/Total (Risk) 112/138 (0.81)   53/90 (0.59)          
#> 8 RR                    1 (reference)    0.73 (0.60-0.88)      
#> 9 RD                    0 (reference)    -0.22 (-0.34 to -0.10)

# Use 'design' as columns (selecting RR and RD only)
table2(design = design1 %>%
                  dplyr::filter(label %in% c("RR", "RD")),
       data = cancer,
       layout = "cols")

#> # A tibble: 2 x 3
#>   sex   RR               RD                    
#>   <chr> <chr>            <chr>                 
#> 1 Men   1 (reference)    0 (reference)         
#> 2 Women 0.73 (0.60-0.88) -0.22 (-0.34 to -0.10)

# Example 2: Survival outcomes (use 'time' and 'event'),
#   with an effect modifier and a confounder
# Set table design
design2 <- tibble::tribble(
  # Elements that vary by row:
  ~label,                       ~stratum, ~confounders, ~type,
  "Overall: Events",            NULL,     "",           "events",
  "  Person-years",             NULL,     "",           "time",
  "  Rate/1000 py (95% CI)",    NULL,     "",           "rate (ci)",
  "  Unadjusted HR (95% CI)",   NULL,     "",           "hr",
  "  Age-adjusted HR (95% CI)", NULL,     "+ age",      "hr",
  "",                           NULL,     "",           "blank",
  "Stratified models",          NULL,     "",           "blank",
  "ECOG PS1 (events/N)",        1,        "",           "events/total",
  "  Unadjusted",               1,        "",           "hr",
  "  Age-adjusted",             1,        "+ age",      "hr",
  "ECOG PS2 (events/N)",        2,        "",           "events/total",
  "  Unadjusted",               2,        "",           "hr",
  "  Age-adjusted",             2,        "+ age",      "hr",
  "",                           NULL,     "",           "",
  "Joint model, age-adj.",      NULL,     "",           "",
  "  ECOG PS1",                 1,        "+ age",      "hr_joint",
  "  ECOG PS2",                 2,        "+ age",      "hr_joint") %>%
  # Elements that are the same for all rows:
  dplyr::mutate(exposure = "sex",
                event = "status",
                time = "time",
                effect_modifier = "ph.ecog")

# Generate table2
table2(design = design2,
       data = cancer %>% dplyr::filter(ph.ecog %in% 1:2))

#> # A tibble: 17 x 3
#>    sex                          Men                     Women                 
#>    <chr>                        <chr>                   <chr>                 
#>  1 "Overall: Events"            "82"                    "44"                  
#>  2 "  Person-years"             "70"                    "59"                  
#>  3 "  Rate/1000 py (95% CI)"    "1164.8 (938.1-1446.3)" "746.7 (555.7-1003.4)"
#>  4 "  Unadjusted HR (95% CI)"   "1 (reference)"         "0.60 (0.41-0.86)"    
#>  5 "  Age-adjusted HR (95% CI)" "1 (reference)"         "0.60 (0.41-0.86)"    
#>  6 ""                           ""                      ""                    
#>  7 "Stratified models"          ""                      ""                    
#>  8 "ECOG PS1 (events/N)"        "54/71"                 "28/42"               
#>  9 "  Unadjusted"               "1 (reference)"         "0.53 (0.33-0.85)"    
#> 10 "  Age-adjusted"             "1 (reference)"         "0.53 (0.33-0.85)"    
#> 11 "ECOG PS2 (events/N)"        "28/29"                 "16/21"               
#> 12 "  Unadjusted"               "1 (reference)"         "0.70 (0.37-1.30)"    
#> 13 "  Age-adjusted"             "1 (reference)"         "0.68 (0.34-1.36)"    
#> 14 ""                           ""                      ""                    
#> 15 "Joint model, age-adj."      ""                      ""                    
#> 16 "  ECOG PS1"                 "1 (reference)"         "0.55 (0.35-0.88)"    
#> 17 "  ECOG PS2"                 "1.54 (0.98-2.44)"      "1.10 (0.62-1.98)"    

# Example 3: Continuous outcomes (use 'outcome' variable);
# request rounding to 1 decimal digit in some cases.
tibble::tribble(
  ~label,                   ~stratum, ~type,
  "Marginal mean (95% CI)", 1:2,      "mean (ci) 1",
  "  Men",                  "Men",    "mean",
  "  Women",                "Women",  "mean",
  "",                       NULL,     "",
  "Stratified model",       NULL,     "",
  "  Men",                  "Men",    "diff 1",
  "  Women",                "Women",  "diff 1",
  "",                       NULL,     "",
  "Joint model",            NULL,     "",
  "  Men",                  "Men",    "diff_joint",
  "  Women",                "Women",  "diff_joint") %>%
  dplyr::mutate(exposure = "ph.ecog",
                outcome = "age",
                effect_modifier = "sex") %>%
  table2(data = cancer %>%
                  dplyr::filter(ph.ecog < 3) %>%
                  dplyr::mutate(ph.ecog = factor(ph.ecog)))

#> # A tibble: 10 x 4
#>    ph.ecog          `0`                  `1`                  `2`               
#>    <chr>            <chr>                <chr>                <chr>             
#>  1 "  Men"          "63.00"              "62.79"              "65.00"           
#>  2 "  Women"        "58.70"              "59.19"              "67.90"           
#>  3 ""               ""                   ""                   ""                
#>  4 "Stratified mod… ""                   ""                   ""                
#>  5 "  Men"          "0 (reference)"      "-0.2 (-3.9 to 3.5)" "2.0 (-2.5 to 6.5…
#>  6 "  Women"        "0 (reference)"      "0.5 (-3.5 to 4.5)"  "9.2 (4.5 to 13.9…
#>  7 ""               ""                   ""                   ""                
#>  8 "Joint model"    ""                   ""                   ""                
#>  9 "  Men"          "0 (reference)"      "-0.21 (-3.75 to 3.… "2.00 (-2.32 to 6…
#> 10 "  Women"        "-4.30 (-8.70 to 0.… "-3.81 (-7.74 to 0.… "4.90 (0.15 to 9.…

# Get formatted output:
if (FALSE) {
table2(design = design2, data = cancer) %>%
  mygt()
}